pyproject.toml
[tool.poetry]
name = "polk-russia-bot"
version = "0.1.0"
description = "The official bot of the Immortal Regiment of Russia..."
authors = ["AnyGlitch <WhyGlitches@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/AnyGlitch/polk-russia-bot"
repository = "https://github.com/AnyGlitch/polk-russia-bot.git"
keywords = ["python", "aiogram"]
include = ["CHANGELOG.md", "LICENSE", "README.md"]

classifiers = [
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Operating System :: POSIX",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = {extras = ["redis"], version = "^3.0.0b", allow-prereleases = true}
aerich = "^0.7.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
pre-commit = "^3.2.2"
mypy = "^1.2.0"
ruff = "^0.0.263"
black = "^23.3.0"

[tool.aerich]
tortoise_orm = "source.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 80
target-version = "py310"
select = [
  "F",
  "E",
  "W",
  "I",
  "UP",
  "B",
  "COM",
  "C4",
  "EXE",
  "G",
  "PIE",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SIM",
  "TID",
  "TCH",
  "ERA",
]

[tool.black]
line-length = 80
target-version = ["py310"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
